+++ python/ray/train/tests/test_minimal.py
@@ -6,7 +6,7 @@ import ray
 import ray.train as train
 from ray.train import Trainer
 from ray.train.backend import BackendConfig, Backend
-from ray.train.callbacks.callback import TrainingCallback
+from ray.train.callbacks import TrainingCallback
 from ray.train.worker_group import WorkerGroup
